Move network transport to netchan
Spdy is no longer the only possible transport for libchan, therefore rename to netchan.
