+++ netchan/README.md
@@ -1,0 +1,3 @@
+This package defines a network transport for libchan using multiplexed streams
+
+Uses https://github.com/dmcgowan/streams
