--- spdy/MAINTAINERS
@@ -1,1 +1,0 @@
-Derek McGowan <derek@docker.com> (@dmcg)
