--- spdy/README.md
@@ -1,8 +1,0 @@
-This package defines a remote transport for libchan services using http2/spdy and tls.
-
-Uses https://github.com/docker/spdystream
-
-Pointers:
-	
-	* Low-level protocol framer: http://code.google.com/p/go.net/spdy
-	* (incomplete) high-level server implementation: https://github.com/shykes/spdy-go
